Add favorites feature with new screen and navigation
Introduce a new favorites feature including a dedicated screen and navigation. Added a new route `/favorites` in `app_routes.dart` and integrated it into the route generator. Created `FavoritesScreen` and `FavoritesBodySection` components for the UI. Added a new asset `favorites.png` and updated `AppAssets` and `AppStrings` to include the new asset and label. Enhanced the home screen by adding a `HomeFavoriteTextSection` widget to navigate to the favorites screen. Improved code formatting in `HomeCharactersListSection` for better readability.