Add internet connectivity check and offline handling
Added the `flutter_offline` package to handle internet connectivity checks in the app. Implemented an offline mode in the `HomeScreen` that displays a connection error message and a Wi-Fi SVG icon when no internet is detected. Updated `AndroidManifest.xml` to include the `INTERNET` permission. Added a new `wifi.svg` asset and updated `AppAssets` and `AppStrings` to support the new functionality. Also, updated `pubspec.yaml` and `pubspec.lock` to include the new dependencies.